1. login.html - connected up the "forgot password" link
2. password_reset.html - this is the form for actually resetting the password

3. request_password_reset.html - this is the publically accessible form, where you ask for a reset email to be sent to your account email

4. __init__.py - updated to connect to gmail using SSL

5. forms.py - added forms which handle the gui+validation for password reset feature

6. models.py - added some fns to User.
a. verify_reset_token(token) -> evaluates the token, and returns a dict, which contains the user id, or the exception which occurred

b. get_reset_token - creates a secure token with a max lifetime, and returns it as a utf-8 string

7. routes.py - added the routes for reset-password feature
a. send_reset_email - handles the actual sending of the reset email, using credentials stored in the dev environment

b. request_password_reset - route which is used to ask for a password reset, which is emailed to the user's account email

b. password_reset - the actual reset step; requires verification, flashes errors that occur